requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
openpyxl
python-dotenv
schedule
requests
selenium
undetected-chromedriver
psycopg2-binary